removed console.log()s
there were a bunch of console logs left in from troubleshooting and building.
cleaned those up, deleted most, left a couple in the big components commented for ease of future troubleshooting.

the only active console.log() [i think] is in App.js as a reminder that the app state is purged from the device while in development.